+++ docs/ChangeLog.md
@@ -24,6 +24,7 @@
 - keymap/vi_nmap: support `shell-expand-line` `#D2145` aa92b42a
 - main: support `bash ble.sh --install` `#D2169` 986d26a3 3801a87e
 - util(stty): support `bleopt term_stty_restore` (requested by TheFantasticWarrior) `#D2170` e64b02b7
+  - util: update workaround of Bash 5.2 `checkwinsize` for `term_stty_restore` (reported by TheFantasticWarrior) `#D2184` xxxxxxxx
 - edit: support `bleopt edit_magic_accept` (requested by pl643, bkerin) `#D2175` 3e9d8907
 - main: support shell variable `BLE_VER` `#D2177` a12dedab
 - edit: support `bleopt edit_magic_accept=verify-syntax` `#D2178` ac84c153